pyproject.toml
[project]
name = "voice-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.14"
dependencies = [
    "edge-tts>=7.2.7",
    "httpx>=0.28.1",
    "kokoro>=0.9.4",
    "livekit-agents[deepgram,openai,silero,turn-detector]>=1.4.5",
    "livekit-plugins-deepgram>=1.4.5",
    "livekit-plugins-google>=1.5.1",
    "livekit-plugins-noise-cancellation>=0.2.5",
    "livekit-plugins-openai>=1.4.5",
    "livekit-plugins-silero>=1.4.5",
    "livekit-plugins-turn-detector>=1.4.5",
    "numpy>=2.4.3",
    "pip>=26.0.1",
    "pypdf>=6.9.2",
    "python-dotenv>=1.2.2",
    "reportlab>=4.4.10",
    "soundfile>=0.13.1",
]
